předěláno na utility object
