+++ lesson-21/index.php
@@ -42,7 +42,7 @@
             }
 
             $user = new User($name, $phone, $email, $message);
-            $fileCreated = new Storage($name, $formData);
+            Storage::save($name, $formData);
 
         } catch (Mail\MailerException $e) {
             $errorCaught = 'Email se nepovedlo odeslat z tohoto důvodu: ' . $e->getMessage();
